moved converter class to converter module (in order to ease integration of build tools other than maven)
