+++ asciidoc-confluence-publisher-converter/pom.xml
@@ -33,6 +33,12 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.sahli.asciidoc.confluence.publisher</groupId>
+            <artifactId>asciidoc-confluence-publisher-client</artifactId>
+            <scope>compile</scope>
+        </dependency>
+
+        <dependency>
             <groupId>commons-lang</groupId>
             <artifactId>commons-lang</artifactId>
             <scope>compile</scope>
